+++ dev-env/script/ruby-dev.sh
@@ -4,18 +4,20 @@ cd $HOME
 . "$HOME/.zshrc"
 RUBY_VERSION=2.4.4
 
-if cat /etc/os-release | grep -sq "Ubuntu" || uname -a | grep -sq "Microsoft"; then
-    sudo apt-get install -y autoconf \
-        bison \
-        build-essential \
-        libssl-dev \
-        libyaml-dev \
-        libreadline-dev \
-        zlib1g-dev \
-        libncurses-dev \
-        libffi-dev \
-        libgdbm3 \
-        libgdbm-dev
+if [ -e /etc/os-release ]; then
+    if cat /etc/os-release | grep -sq "Ubuntu" || uname -a | grep -sq "Microsoft"; then
+        sudo apt-get install -y autoconf \
+            bison \
+            build-essential \
+            libssl-dev \
+            libyaml-dev \
+            libreadline-dev \
+            zlib1g-dev \
+            libncurses-dev \
+            libffi-dev \
+            libgdbm3 \
+            libgdbm-dev
+    fi
 elif uname -a | grep -sq "Darwin"; then
     brew install openssl libyaml libffi
 fi
